+++ Assets/lilToon/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.5] - 2022-07-xx
 ### Added
 - LightMode override
+- `Tiling & Offset` and `Blend Mode (Add / Subtract)` for alpha mask
 
 ### Fixed
 - Fixed error in VRChat World SDK